# docs/swagger.yaml
openapi: 3.0.0
info:
  title: Go HRM Microservices API
  description: API documentation for Go HRM system with microservices architecture
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
servers:
  - url: http://localhost:8081
    description: User Service
  - url: http://localhost:8082
    description: Company Service
  - url: http://localhost:8083
    description: Employee Service
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
        fullname:
          type: string
        email:
          type: string
        status:
          type: integer
        datelastlogin:
          type: integer
    Company:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        domain:
          type: string
        owner:
          type: integer
        status:
          type: integer
    Employee:
      type: object
      properties:
        id:
          type: integer
        internal_id:
          type: string
        email:
          type: string
        status:
          type: integer
        user_id:
          type: integer
        company_id:
          type: integer
    LoginRequest:
      type: object
      required:
        - email
        - password
        - hostname
        - platform
        - version
      properties:
        email:
          type: string
          format: email
          example: "admin@gmail.com"
        password:
          type: string
          example: "admin123"
        hostname:
          type: string
          example: "local.localhost"
        platform:
          type: string
          example: "web"
        version:
          type: string
          example: "1.0.0"
        internalid:
          type: string
          description: "Optional internal ID"
    LoginResponse:
      type: object
      properties:
        jwt:
          type: string
        refresh_token:
          type: string
        expires_at:
          type: integer
        user:
          $ref: '#/components/schemas/User'
        company:
          $ref: '#/components/schemas/Company'
        employee:
          $ref: '#/components/schemas/Employee'
        role:
          type: string
...
    RefreshTokenRequest:
      type: object
      required:
        - refresh_token
      properties:
        refresh_token:
          type: string
          description: "Refresh token received from login"
    RefreshTokenResponse:
      type: object
      properties:
        jwt:
          type: string
        refresh_token:
          type: string
        expires_at:
          type: integer
paths:
...
  /api/v1/auth/login:
    post:
      tags:
        - Authentication
      summary: User login
      description: Authenticate user and return JWT tokens
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          description: Invalid request
        '401':
          description: Invalid credentials
  /api/v1/auth/logout:
    post:
      tags:
        - Authentication
      summary: User logout
      description: Logout user and invalidate session
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Logout successful
        '401':
          description: Unauthorized
  /api/v1/auth/refresh:
    post:
      tags:
        - Authentication
      summary: Refresh access token
      description: Get new access token using refresh token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefreshTokenResponse'
        '401':
          description: Invalid refresh token
...